+++ CHANGELOG.rst
@@ -14,6 +14,10 @@ Added
 -----
 - ``row`` and ``col`` properties to ``CrystalMap`` giving the row and column coordinate
   of each map point given by ``CrystalMap.shape``.
+- ``Rotation`` class methods ``from_neo_euler()``, ``from_axes_angles()``,
+  ``from_euler()``, ``from_matrix()``, ``random()`` and ``identity()`` and methods
+  ``to_euler()`` and ``to_matrix()`` are now available from the ``Quaternion`` class as
+  well.
 
 Changed
 -------